use smaller datatypes for the tables where possible

+++ wmapro/wma3data.h
@@ -405,7 +405,7 @@ static const uint8_t ff_wma3_huff_rlc_bits[FF_WMA3_HUFF_RLC_SIZE] = {
 };
 
 
-static const short ff_wma3_run_0[] = {
+static const uint8_t ff_wma3_run_0[] = {
      0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16, 17,
     18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35,
     36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53,
@@ -423,7 +423,7 @@ static const short ff_wma3_run_0[] = {
      0,  0,  0,  0,
 };
 
-static const short ff_wma3_level_0[] = {
+static const uint8_t ff_wma3_level_0[] = {
      1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
      1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
      1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
@@ -442,7 +442,7 @@ static const short ff_wma3_level_0[] = {
 };
 
 
-static const short ff_wma3_run_1[] = {
+static const uint8_t ff_wma3_run_1[] = {
       0,   1,   2,   3,   4,   5,   6,   7,   8,   9,  10,  11,  12,  13,
      14,  15,  16,  17,  18,  19,  20,  21,  22,  23,  24,  25,  26,  27,
      28,  29,  30,  31,  32,  33,  34,  35,  36,  37,  38,  39,  40,  41,
@@ -465,7 +465,7 @@ static const short ff_wma3_run_1[] = {
       1,   0,   1,   0,   0,   0,
 };
 
-static const short ff_wma3_level_1[] = {
+static const uint8_t ff_wma3_level_1[] = {
       1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
       1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
       1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
